Force some errors

+++ app/lib/actions.ts
@@ -6,29 +6,37 @@ import { CartProduct } from '@/models/CartProduct';
 import { redirect } from 'next/navigation';
 
 export async function updateQuantityProductInCart(id: number, quantity: number) {
-	try {
-		const response = await fetch(`http://localhost:3001/carts/${id}${id === 12 ? 'error' : ''}`, {
+	let response: Response | undefined = undefined;
+	if (id === 12) {
+		response = await fetch('https://httpstat.us/500?sleep=2000');
+	} else {
+		response = await fetch(`http://localhost:3001/carts/${id}`, {
+			cache: 'no-cache',
 			method: 'PATCH',
 			headers: {
 				'Content-Type': 'application/json'
 			},
 			body: JSON.stringify({ quantity })
 		});
-
-		if (!response.ok) {
-			return { error: 'ERROR_PATCHING' };
-		}
-	} catch (e) {
-		console.error('error server action', e);
 	}
 
-	revalidatePath('/carrello', 'page');
+	if (response.status === 500) {
+		return { error: 'ERROR_PATCHING' };
+	}
+
+	revalidatePath('/carrello');
 }
 
 export async function addProductToCart(product: Product) {
-	const productPresentResponse = await fetch(`http://localhost:3001/carts/${product.id}`);
-	if (productPresentResponse.status.toString().split('')[0] === '5') {
-		throw new Error('Error fetching product');
+	let productPresentResponse: Response | undefined = undefined;
+	if (product.id === 4) {
+		productPresentResponse = await fetch('https://httpstat.us/500?sleep=2000');
+	} else {
+		productPresentResponse = await fetch(`http://localhost:3001/carts/${product.id}`, { cache: 'no-cache' });
+	}
+
+	if (productPresentResponse.status === 500) {
+		return { error: 'ERROR_ADDING' };
 	}
 
 	let productInCart: CartProduct | undefined = undefined;
@@ -37,9 +45,24 @@ export async function addProductToCart(product: Product) {
 	}
 
 	if (productInCart) {
-		return updateQuantityProductInCart(productInCart.id, productInCart.quantity + 1);
+		const response = await fetch(`http://localhost:3001/carts/${productInCart.id}`, {
+			cache: 'no-cache',
+			method: 'PATCH',
+			headers: {
+				'Content-Type': 'application/json'
+			},
+			body: JSON.stringify({ quantity: productInCart.quantity + 1 })
+		});
+
+		if (!response.ok) {
+			return { error: 'ERROR_PATCHING' };
+		}
+
+		revalidatePath('/carrello', 'page');
+		redirect('/carrello');
 	} else {
 		const response = await fetch(`http://localhost:3001/carts`, {
+			cache: 'no-cache',
 			method: 'POST',
 			headers: {
 				'Content-Type': 'application/json'
@@ -48,7 +71,7 @@ export async function addProductToCart(product: Product) {
 		});
 
 		if (!response.ok) {
-			throw new Error(`Error adding product to cart. Error ${JSON.stringify(response)}`);
+			return { error: 'ERROR_ADDING_PRODUCT' };
 		}
 
 		revalidatePath('/carrello', 'page');
@@ -57,12 +80,17 @@ export async function addProductToCart(product: Product) {
 }
 
 export async function removeProductFromCart(id: number) {
-	const response = await fetch(`http://localhost:3001/carts/${id}`, {
-		method: 'DELETE'
-	});
+	let response: Response | undefined = undefined;
+	if (id === 6) {
+		response = await fetch('https://httpstat.us/500?sleep=2000');
+	} else {
+		response = await fetch(`http://localhost:3001/carts/${id}`, {
+			method: 'DELETE'
+		});
+	}
 
 	if (!response.ok) {
-		throw new Error(`Error deleting product with id ${id}. Error ${JSON.stringify(response)}`);
+		return { error: 'ERROR_DELETING' };
 	}
 
 	revalidatePath('/carrello', 'page');
